% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{check_colours}
\alias{check_colours}
\title{Check whether you can map your data to a set of colours and guess the appropriate group based on overlap}
\usage{
check_colours(groups)
}
\arguments{
\item{groups}{}
}
\value{
The name of the group that is the best match for your labels
}
\description{
Check whether you can map your data to a set of colours and guess the appropriate group based on overlap
}
\examples{

my_groups = c("ABC","GCB","U")
check_colours(my_groups)

}
