% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_colours}
\alias{get_colours}
\title{Get a set of standard colours as hex codes}
\usage{
get_colours(col_category, col_group, return_format = "vector")
}
\arguments{
\item{col_category}{}
}
\value{
Depending on return_format, a vector or list or data frame
}
\description{
Get a set of standard colours as hex codes
}
\examples{

col_vector = get_colours(col_group="BL-genetic")

}
