% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{show_colours}
\alias{show_colours}
\title{Show all the colour codes in a table}
\usage{
show_colours(col_category, col_group)
}
\value{
Nothing
}
\description{
Show all the colour codes in a table
}
\examples{
show_colours()
show_colours(col_category = "subgroup")
show_colours(col_group = "BL-genetic")
}
