% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{summarise_colours}
\alias{summarise_colours}
\title{Summarize all the colour codes in a table after applying some or no filtering criteria}
\usage{
summarise_colours(col_category, col_group)
}
\value{
Nothing
}
\description{
Summarize all the colour codes in a table after applying some or no filtering criteria
}
\examples{

summarise_colours()
summarise_colours(col_category="subgroup")
summarise_colours(col_category="patient")


}
